update templates-namespace
The 1.0.2 version went to both the original namespace
https://github.com/tailscale/codespace/pkgs/container/codespace%2Ftailscale

and to the new, and intended, namespace:
https://github.com/tailscale/codespace/pkgs/container/codespace

Most importantly, the 1.0.2 tag only appears in the original namespace,
and we want it to appear in the new namespace.

Add a templates-namespace declaration, maybe the original namespace is
being used because of the template.

Signed-off-by: Denton Gentry <dgentry@tailscale.com>